el script update

+++ src/pyscripts/pyscript_runner.py
@@ -82,7 +82,7 @@ class RemoteResource(object):
     def read(self, verbose=True):
         resp = self.handle.ask('Read {}'.format(self.name), verbose=verbose)
         if resp is not None:
-            return int(resp)
+            return float(resp)
 
     def isSet(self):
         resp = self.read()
